com.irlballers.api/src/games/serializers.py
from rest_framework import serializers
from .models import *

class GameSerializer(serializers.ModelSerializer):
    home_team = serializers.StringRelatedField()
    away_team = serializers.StringRelatedField()

    class Meta:
        model = Game
        fields = '__all__'

