com.irlballers.api/src/requirements.txt
amqp==5.3.1
asgiref==3.11.1
billiard==4.2.4
celery==5.6.3
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.3
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
Django==6.0.3
django-celery==3.1.17
django-celery-beat==2.9.0
django-cors-headers==4.9.0
django-debug-toolbar==6.2.0
django-timezone-field==7.2.1
djangorestframework==3.17.1
idna==3.11
kombu==5.6.2
nba_api==1.11.4
numpy==2.4.4
packaging==26.1
pandas==3.0.1
pillow==12.1.1
prompt_toolkit==3.0.52
psycopg-binary==3.3.3
psycopg2-binary==2.9.11
python-crontab==3.3.0
python-dateutil==2.9.0.post0
redis==6.4.0
requests==2.33.1
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
